Create Get_Node_Value
Given a pointer to the head of a linked list and a specific position, determine the data value at that position. Count backwards from the tail node. The tail is at postion 0, its parent is at 1 and so on.

Example
head refers to 3->2->1->NULL
positionFromTail=2
Each of the data values matches its distance from the tail. The value 2 is at the desired position.

Function Description
Complete the getNode function in the editor below.


getNode has the following parameters:
1)SinglyLinkedListNode pointer head: refers to the head of the list
2)int positionFromTail: the item to retrieve
Returns
int: the value at the desired position

Input Format
The first line contains an integer t , the number of test cases.

Each test case has the following format:
The first line contains an integer n, the number of elements in the linked list.
The next n lines contains an integer, the data value for an element of the linked list.
The last line contains an integer positionFromTail , the position from the tail to retrieve the value of.